add helpers for some util functions
