+++ src/lib/index.ts
@@ -1,3 +1,4 @@
 export * from './controls';
 export * from './constants';
+export * from './helpers';
 export * from './interfaces';
